scripts/requirements.txt
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.108.0
openai>=1.12.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
